Found a mini bug when making a new Access Token
Turns out that when making a new access token I was passing the contents
as the basic info from the database, when I should also include the
role.
So I just did the same thing as in the ServiceLayer/LogInLogic file, and
hard coded them just in case.
Also, corrected a typo with the function name `generateAccesToken`,
changed it to `generateAccessToken`